refactor!: rename the `ws` folder to `gateway` (#2037)
