+++ gateway/gateway_manager.ts
