+++ gateway/heartbeat.ts
