+++ gateway/processGatewayQueue.ts
